Adjusted react demo to make better use of react internals
